pyproject.toml
[project]
name = "tg-toolkit"
version = "0.1.0"
description = "Minimal Telegram technical toolkit bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.27.0",
    "python-dotenv>=1.1.1",
]
